Multiple changes
- Updated sources to use types from types.rs
- Created function to read map data json from maps/ and load it into the game.
- Consts move from f32 to i32

+++ src/consts.rs
@@ -1,4 +1,4 @@
-pub static DEFAULT_BUTTON_WIDTH: f32 = 100_f32;
-pub static DEFAULT_BUTTON_HEIGHT: f32 = 50_f32;
+pub static DEFAULT_BUTTON_WIDTH: i32 = 100;
+pub static DEFAULT_BUTTON_HEIGHT: i32 = 50;
 pub static SCREEN_WIDTH: i32 = 1280;
 pub static SCREEN_HEIGHT: i32 = 720;
